Untangle chaiscript_engine from the rest of it
