+++ include/chaiscript/chaiscript_stdlib.hpp
@@ -14,11 +14,15 @@
 #include <vector>
 
 #include "chaiscript_defines.hpp"
-#include "dispatchkit/dispatchkit.hpp"
+#include "language/chaiscript_common.hpp"
+
+//#include "dispatchkit/dispatchkit.hpp"
+#include "dispatchkit/operators.hpp"
 #include "dispatchkit/bootstrap.hpp"
 #include "dispatchkit/bootstrap_stl.hpp"
-#include "dispatchkit/boxed_value.hpp"
+//#include "dispatchkit/boxed_value.hpp"
 #include "language/chaiscript_prelude.hpp"
+#include "dispatchkit/register_function.hpp"
 #include "utility/json_wrap.hpp"
 
 #ifndef CHAISCRIPT_NO_THREADS
@@ -38,15 +42,15 @@ namespace chaiscript
 
       static ModulePtr library()
       {
-        using namespace bootstrap;
+ //       using namespace bootstrap;
 
         auto lib = std::make_shared<Module>();
-        Bootstrap::bootstrap(*lib);
+        bootstrap::Bootstrap::bootstrap(*lib);
 
-        standard_library::vector_type<std::vector<Boxed_Value> >("Vector", *lib);
-        standard_library::string_type<std::string>("string", *lib);
-        standard_library::map_type<std::map<std::string, Boxed_Value> >("Map", *lib);
-        standard_library::pair_type<std::pair<Boxed_Value, Boxed_Value > >("Pair", *lib);
+        bootstrap::standard_library::vector_type<std::vector<Boxed_Value> >("Vector", *lib);
+        bootstrap::standard_library::string_type<std::string>("string", *lib);
+        bootstrap::standard_library::map_type<std::map<std::string, Boxed_Value> >("Map", *lib);
+        bootstrap::standard_library::pair_type<std::pair<Boxed_Value, Boxed_Value > >("Pair", *lib);
 
 #ifndef CHAISCRIPT_NO_THREADS
         standard_library::future_type<std::future<chaiscript::Boxed_Value>>("future", *lib);
